# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {"title": "Glimse daemon API", "version": "1.0.0", "description": "Versioned publication, artifact delivery, lifecycle, read, and live-feed contract. Runtime access may use Bearer or browser-session credentials, or trusted-proxy authorization based only on the immediate TCP peer; health remains public."},
  "x-v1-routing-errors": {"unknown_route": {"status": 404, "code": "api_route_not_found"}, "wrong_method": {"status": 405, "code": "method_not_allowed"}},
  "security": [{"BearerToken": []}, {"BrowserSession": []}],
  "paths": {
    "/api/v1/health": {"get": {"operationId": "health", "security": [], "responses": {"200": {"description": "Healthy", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Health"}}}}}}},
    "/api/v1/auth/session": {
      "post": {"operationId": "createBrowserSession", "security": [], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrowserSessionRequest"}}}}, "responses": {"204": {"description": "Browser session cookie issued"}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}}},
      "delete": {"operationId": "deleteBrowserSession", "responses": {"204": {"description": "Browser session invalidated and cookie expired"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}}}
    },
    "/api/v1/sessions": {"post": {"operationId": "resolveSession", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResolveSessionRequest"}}}}, "responses": {"200": {"description": "Resolved session", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Session"}}}}, "400": {"$ref": "#/components/responses/Error"}, "422": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/sessions/{public_id}": {
      "parameters": [{"$ref": "#/components/parameters/PublicId"}],
      "get": {"operationId": "getSession", "responses": {"200": {"description": "Session", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Session"}}}}, "404": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}},
      "delete": {"operationId": "closeSession", "responses": {"200": {"description": "Idempotent lifecycle counts", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LifecycleReport"}}}}, "500": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}
    },
    "/api/v1/sessions/{public_id}/posts": {"parameters": [{"$ref": "#/components/parameters/PublicId"}, {"$ref": "#/components/parameters/Limit"}, {"$ref": "#/components/parameters/Cursor"}], "get": {"operationId": "listSessionPosts", "responses": {"200": {"$ref": "#/components/responses/PostPage"}, "400": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/sessions/{public_id}/posts/events": {"parameters": [{"$ref": "#/components/parameters/PublicId"}, {"$ref": "#/components/parameters/LastEventId"}], "get": {"operationId": "sessionPostEvents", "description": "Session-scoped SSE. Post IDs are event IDs. Replay is limited to 100 posts; reset signals require latest-page reconciliation.", "responses": {"200": {"$ref": "#/components/responses/EventStream"}, "400": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/sessions/{public_id}/heartbeat": {"parameters": [{"$ref": "#/components/parameters/PublicId"}], "post": {"operationId": "heartbeat", "description": "Records visible-session activity using the daemon's wall-clock Unix time. The request has no body.", "responses": {"200": {"description": "Monotonic activity result using server-owned time", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActivityReport"}}}}, "400": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/projects/{project_id}/posts": {"parameters": [{"$ref": "#/components/parameters/ProjectId"}, {"$ref": "#/components/parameters/Limit"}, {"$ref": "#/components/parameters/Cursor"}], "get": {"operationId": "listProjectPosts", "responses": {"200": {"$ref": "#/components/responses/PostPage"}, "400": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/projects/{project_id}/posts/events": {"parameters": [{"$ref": "#/components/parameters/ProjectId"}, {"$ref": "#/components/parameters/LastEventId"}], "get": {"operationId": "projectPostEvents", "description": "Project-scoped SSE with bounded durable replay and session-closure notifications.", "responses": {"200": {"$ref": "#/components/responses/EventStream"}, "400": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/posts/events": {"get": {"operationId": "globalPostEvents", "description": "Global SSE with bounded durable replay and session-closure notifications.", "parameters": [{"$ref": "#/components/parameters/LastEventId"}], "responses": {"200": {"$ref": "#/components/responses/EventStream"}, "400": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/posts": {
      "get": {"operationId": "listGlobalPosts", "parameters": [{"$ref": "#/components/parameters/Limit"}, {"$ref": "#/components/parameters/Cursor"}], "responses": {"200": {"$ref": "#/components/responses/PostPage"}, "400": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}},
      "post": {"operationId": "publishPost", "description": "Streams one atomic publication. The first part must be the UTF-8 JSON `manifest` (maximum 64 KiB). At most 256 uniquely named declared artifact parts may follow, in any arrival order. Artifact fields are streamed and client filenames and MIME declarations are not authoritative.", "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"type": "object", "required": ["manifest"], "properties": {"manifest": {"$ref": "#/components/schemas/PublicationManifest"}}, "additionalProperties": {"type": "string", "format": "binary"}, "x-part-order": "manifest must be first; every declared byte part follows exactly once"}, "encoding": {"manifest": {"contentType": "application/json"}}}}}, "responses": {"201": {"description": "Atomically published immutable post and resolved session", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicationResponse"}}}}, "400": {"$ref": "#/components/responses/Error"}, "413": {"$ref": "#/components/responses/Error"}, "422": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}, "500": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}, "507": {"$ref": "#/components/responses/Error"}}}
    },
    "/api/v1/posts/{post_id}": {"parameters": [{"$ref": "#/components/parameters/PostId"}], "get": {"operationId": "getPost", "responses": {"200": {"description": "Immutable post", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Post"}}}}, "400": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/posts/{post_id}/files/{position}/html-capability": {"parameters": [{"$ref": "#/components/parameters/PostId"}, {"$ref": "#/components/parameters/FilePosition"}], "post": {"operationId": "createHtmlCapability", "description": "Issues a bounded support-path prefix for one HTML artifact. In local mode the ordinary authenticated support prefix is returned.", "responses": {"200": {"description": "Scoped support prefix", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HtmlCapability"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/posts/{post_id}/files/{position}/content": {"parameters": [{"$ref": "#/components/parameters/PostId"}, {"$ref": "#/components/parameters/FilePosition"}], "get": {"operationId": "getVisibleArtifact", "description": "Streams the associated immutable visible file. One bytes range is accepted; malformed, multiple, or unsatisfiable ranges return 416 with an empty body.", "responses": {"200": {"description": "Complete artifact with its classified dynamic Content-Type", "content": {"*/*": {"schema": {"type": "string", "format": "binary"}}}}, "206": {"description": "Selected artifact byte range with its classified dynamic Content-Type", "content": {"*/*": {"schema": {"type": "string", "format": "binary"}}}}, "400": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "416": {"description": "Range not satisfiable; empty body and Content-Range bytes */length"}, "500": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}, "head": {"operationId": "headVisibleArtifact", "responses": {"200": {"description": "Complete artifact headers with the classified dynamic Content-Type", "content": {"*/*": {"schema": {"type": "string", "format": "binary"}}}}, "206": {"description": "Selected range headers", "content": {"*/*": {"schema": {"type": "string", "format": "binary"}}}}, "400": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "416": {"description": "Range not satisfiable"}, "500": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/posts/{post_id}/files/{position}/support/{asset_path}": {"parameters": [{"$ref": "#/components/parameters/PostId"}, {"$ref": "#/components/parameters/FilePosition"}, {"$ref": "#/components/parameters/AssetPath"}], "get": {"operationId": "getSupportArtifact", "description": "Streams the support asset associated with this exact post, visible-file position, and stored relative path. Encoded separators, traversal segments, duplicate slashes, control characters, and malformed encoding are rejected.", "responses": {"200": {"description": "Complete support artifact with its bounded-classified dynamic Content-Type", "content": {"*/*": {"schema": {"type": "string", "format": "binary"}}}}, "206": {"description": "Selected support artifact byte range", "content": {"*/*": {"schema": {"type": "string", "format": "binary"}}}}, "400": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "416": {"description": "Range not satisfiable"}, "500": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}, "head": {"operationId": "headSupportArtifact", "responses": {"200": {"description": "Complete support artifact headers with dynamic Content-Type", "content": {"*/*": {"schema": {"type": "string", "format": "binary"}}}}, "206": {"description": "Selected support range headers", "content": {"*/*": {"schema": {"type": "string", "format": "binary"}}}}, "400": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "416": {"description": "Range not satisfiable"}, "500": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}}
  },
  "components": {
    "securitySchemes": {
      "BearerToken": {"type": "http", "scheme": "bearer", "bearerFormat": "64 lowercase hexadecimal characters"},
      "BrowserSession": {"type": "apiKey", "in": "cookie", "name": "glim_session"}
    },
    "parameters": {
      "PublicId": {"name": "public_id", "in": "path", "required": true, "schema": {"type": "string"}},
      "ProjectId": {"name": "project_id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64", "minimum": 1}},
      "PostId": {"name": "post_id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64", "minimum": 1}},
      "FilePosition": {"name": "position", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 0}},
      "AssetPath": {"name": "asset_path", "in": "path", "required": true, "description": "Stored relative support path, including nested segments.", "schema": {"type": "string"}},
      "Limit": {"name": "limit", "in": "query", "required": false, "description": "Page size; defaults to 20 and cannot exceed 100.", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 20}},
      "Cursor": {"name": "cursor", "in": "query", "required": false, "description": "Exclusive published_at:post_id cursor from next_cursor. Ordering is published_at DESC, then post ID DESC.", "schema": {"type": "string", "pattern": "^-?[0-9]+:[1-9][0-9]*$"}},
      "LastEventId": {"name": "Last-Event-ID", "in": "header", "required": false, "description": "Last processed positive post event ID.", "schema": {"type": "integer", "format": "int64", "minimum": 1}}
    },
    "responses": {
      "Error": {"description": "Stable error envelope", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "PostPage": {"description": "Bounded reverse-publication-order page", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostPage"}}}},
      "EventStream": {"description": "Server-sent post, session-closed, and reset events with 15-second keepalive comments", "content": {"text/event-stream": {"schema": {"type": "string"}}}},
      "Artifact": {"description": "Streaming immutable bytes with a dynamically selected Content-Type, Content-Length, Accept-Ranges, Content-Disposition, private immutable caching, nosniff, and restrictive CSP headers", "content": {"*/*": {"schema": {"type": "string", "format": "binary"}}}}
    },
    "schemas": {
      "Health": {"type": "object", "additionalProperties": false, "required": ["ok", "version"], "properties": {"ok": {"type": "boolean"}, "version": {"type": "string"}}},
      "BrowserSessionRequest": {"type": "object", "additionalProperties": false, "required": ["token"], "properties": {"token": {"type": "string", "pattern": "^[0-9a-f]{64}$"}}},
      "HtmlCapability": {"type": "object", "additionalProperties": false, "required": ["path_prefix", "expires_in_seconds"], "properties": {"path_prefix": {"type": "string"}, "expires_in_seconds": {"type": "integer", "minimum": 1}}},
      "ResolveSessionRequest": {"type": "object", "additionalProperties": false, "required": ["integration_namespace", "external_key", "project_label", "working_directory"], "properties": {"integration_namespace": {"type": "string"}, "external_key": {"type": "string"}, "project_label": {"type": "string"}, "working_directory": {"type": "string", "description": "Identity metadata only; never read as a source path."}}},
      "PublicationManifest": {"type": "object", "additionalProperties": false, "required": ["integration_namespace", "external_key", "project_label", "working_directory", "title", "commentary", "files"], "properties": {"integration_namespace": {"type": "string"}, "external_key": {"type": "string"}, "project_label": {"type": "string"}, "working_directory": {"type": "string", "description": "Identity metadata only; never read as a source path."}, "title": {"type": "string"}, "commentary": {"type": "string"}, "predecessor_post_id": {"type": ["integer", "null"], "format": "int64"}, "git": {"anyOf": [{"$ref": "#/components/schemas/GitProvenance"}, {"type": "null"}], "description": "Inert publication metadata. The daemon never executes Git or reads this path."}, "files": {"type": "array", "minItems": 1, "maxItems": 256, "items": {"$ref": "#/components/schemas/PublicationManifestFile"}}}},
      "PublicationManifestFile": {"type": "object", "additionalProperties": false, "required": ["part", "filename"], "properties": {"part": {"type": "string", "minLength": 1}, "filename": {"type": "string"}, "caption": {"type": ["string", "null"]}, "media_type": {"type": ["string", "null"], "description": "Optional authoritative declaration validated against filename and a bounded byte prefix."}, "support_assets": {"type": "array", "default": [], "maxItems": 255, "items": {"$ref": "#/components/schemas/PublicationManifestSupportAsset"}}}},
      "PublicationManifestSupportAsset": {"type": "object", "additionalProperties": false, "required": ["part", "relative_path"], "properties": {"part": {"type": "string", "minLength": 1}, "relative_path": {"type": "string"}}},
      "PublicationResponse": {"type": "object", "additionalProperties": false, "required": ["session", "post"], "properties": {"session": {"$ref": "#/components/schemas/Session"}, "post": {"$ref": "#/components/schemas/Post"}}},
      "GitProvenance": {"type": "object", "additionalProperties": false, "required": ["root", "branch", "commit"], "properties": {"root": {"type": "string", "minLength": 1, "maxLength": 4096, "pattern": "^/[^\\u0000-\\u001F\\u007F]*$"}, "branch": {"type": ["string", "null"], "minLength": 1, "maxLength": 1024, "pattern": "^[^\\u0000-\\u001F\\u007F]+$"}, "commit": {"type": ["string", "null"], "pattern": "^(?:[0-9A-Fa-f]{40}|[0-9A-Fa-f]{64})$"}}},
      "Project": {"type": "object", "required": ["id", "label", "working_directory"], "properties": {"id": {"type": "integer", "format": "int64"}, "label": {"type": "string"}, "working_directory": {"type": "string"}}},
      "Session": {"type": "object", "required": ["id", "public_id", "integration_namespace", "external_key", "project", "created_at", "last_activity_at"], "properties": {"id": {"type": "integer", "format": "int64"}, "public_id": {"type": "string"}, "integration_namespace": {"type": "string"}, "external_key": {"type": "string"}, "project": {"$ref": "#/components/schemas/Project"}, "created_at": {"type": "integer", "format": "int64"}, "last_activity_at": {"type": "integer", "format": "int64"}}},
      "Blob": {"type": "object", "required": ["hash", "byte_size"], "properties": {"hash": {"type": "string", "pattern": "^[0-9a-f]{64}$"}, "byte_size": {"type": "integer", "format": "uint64", "minimum": 0}}},
      "SupportAsset": {"type": "object", "required": ["relative_path", "blob"], "properties": {"relative_path": {"type": "string"}, "blob": {"$ref": "#/components/schemas/Blob"}}},
      "PostFile": {"type": "object", "additionalProperties": false, "required": ["position", "filename", "caption", "media_type", "renderer", "blob", "support_assets"], "properties": {"position": {"type": "integer", "minimum": 0}, "filename": {"type": "string"}, "caption": {"type": ["string", "null"]}, "media_type": {"type": "string"}, "renderer": {"type": "string", "enum": ["image", "svg", "pdf", "video", "audio", "markdown", "text", "json", "csv", "html", "download"]}, "blob": {"$ref": "#/components/schemas/Blob"}, "support_assets": {"type": "array", "items": {"$ref": "#/components/schemas/SupportAsset"}}}},
      "Post": {"type": "object", "required": ["id", "session_id", "session_public_id", "title", "commentary", "predecessor_post_id", "published_at", "git", "files"], "properties": {"id": {"type": "integer", "format": "int64"}, "session_id": {"type": "integer", "format": "int64"}, "session_public_id": {"type": "string"}, "title": {"type": "string"}, "commentary": {"type": "string"}, "predecessor_post_id": {"type": ["integer", "null"], "format": "int64"}, "published_at": {"type": "integer", "format": "int64"}, "git": {"anyOf": [{"$ref": "#/components/schemas/GitProvenance"}, {"type": "null"}]}, "files": {"type": "array", "items": {"$ref": "#/components/schemas/PostFile"}}}},
      "PostPage": {"type": "object", "required": ["posts", "next_cursor"], "properties": {"posts": {"type": "array", "items": {"$ref": "#/components/schemas/Post"}}, "next_cursor": {"type": ["string", "null"]}}},
      "ActivityReport": {"type": "object", "required": ["updated", "last_activity_at"], "properties": {"updated": {"type": "boolean"}, "last_activity_at": {"type": "integer", "format": "int64"}}},
      "LifecycleReport": {"type": "object", "required": ["sessions_deleted", "projects_deleted", "posts_deleted", "post_files_deleted", "support_assets_deleted", "blob_references_deleted", "blobs_queued", "blobs_deleted"], "properties": {"sessions_deleted": {"type": "integer", "minimum": 0}, "projects_deleted": {"type": "integer", "minimum": 0}, "posts_deleted": {"type": "integer", "minimum": 0}, "post_files_deleted": {"type": "integer", "minimum": 0}, "support_assets_deleted": {"type": "integer", "minimum": 0}, "blob_references_deleted": {"type": "integer", "minimum": 0}, "blobs_queued": {"type": "integer", "minimum": 0}, "blobs_deleted": {"type": "integer", "minimum": 0}}},
      "ErrorEnvelope": {"type": "object", "additionalProperties": false, "required": ["error"], "properties": {"error": {"type": "object", "additionalProperties": false, "required": ["code", "message", "details"], "properties": {"code": {"type": "string", "pattern": "^[a-z0-9_]+$", "enum": ["authentication_required", "proxy_authorization_required", "authentication_not_configured", "invalid_credentials", "origin_rejected", "malformed_json", "malformed_query", "malformed_path", "malformed_last_event_id", "unexpected_request_body", "malformed_multipart", "multipart_stream_error", "manifest_must_be_first", "manifest_too_large", "manifest_not_utf8", "malformed_manifest", "manifest_complexity_exceeded", "invalid_part_name", "duplicate_manifest", "duplicate_part", "unexpected_part", "missing_part", "duplicate_support_path", "artifact_classification_failed", "artifact_not_found", "invalid_page_limit", "invalid_page_cursor", "validation_failed", "session_not_found", "project_not_found", "post_not_found", "api_route_not_found", "method_not_allowed", "revision_conflict", "storage_constraint_conflict", "upload_limit_exceeded", "storage_limit_exceeded", "storage_integrity_error", "storage_io_error", "database_error", "internal_error", "storage_unavailable"]}, "message": {"type": "string"}, "details": {"type": "object"}}}}}
    }
  }
}